blogproject/blogproject/requirements.txt
asgiref==3.7.2
Django==4.2.2
graphql-core==3.2.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.186.1
strawberry-graphql-django==0.9.5
typing_extensions==4.6.3
